backend/requirements.txt
Django==5.1
djangorestframework==3.15.2
mysqlclient==2.1.1